# gen/http/openapi.json
{"swagger":"2.0","info":{"title":"Co-op Bank HTTP Listeners","description":"HTTP service for interacting with callbacks supporting Co-op Bank APIs","termsOfService":"https://github.com/wondenge/coop-notifiers/blob/master/LICENSE","contact":{"name":"William Ondenge","email":"ondengew@gmail.com","url":"https://www.ondenge.me"},"license":{"name":"Apache License","url":"https://github.com/wondenge/coop-notifiers/blob/master/LICENSE"},"version":"1.0"},"host":"localhost:3000","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/callbacks/coop-bank/FundsTransfer/External/A2A/PesaLink/1.0.0":{"post":{"tags":["connect"],"summary":"PesaLinkSendToAccount connect","description":"PesaLinkSendToAccount Early Hints Callback Request","operationId":"connect#PesaLinkSendToAccount","parameters":[{"name":"PesaLinkSendToAccountRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/ConnectPesaLinkSendToAccountRequestBody","required":["MessageReference","MessageDateTime","MessageCode","MessageDescription","Source","Destinations"]}}],"responses":{"200":{"description":"OK response.","schema":{"type":"string"}},"207":{"description":"Multi-Status response.","schema":{"$ref":"#/definitions/ConnectPesaLinkSendToAccountFullFailureResponseBody"}}},"schemes":["http"]}},"/callbacks/coop-bank/FundsTransfer/External/A2M/Mpesa/v1.0.0":{"post":{"tags":["connect"],"summary":"SendToMPesa connect","description":"SendToMPesa Early Hints Callback Request","operationId":"connect#SendToMPesa","parameters":[{"name":"SendToMPesaRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/ConnectSendToMPesaRequestBody","required":["MessageReference","MessageDateTime","MessageCode","MessageDescription","Source","Destinations"]}}],"responses":{"200":{"description":"OK response.","schema":{"type":"string"}},"207":{"description":"Multi-Status response.","schema":{"$ref":"#/definitions/ConnectSendToMPesaFullFailureResponseBody"}}},"schemes":["http"]}},"/callbacks/coop-bank/FundsTransfer/External/A2M/PesaLink/1.0.0":{"post":{"tags":["connect"],"summary":"PesaLinkSendToPhone connect","description":"PesaLinkSendToPhone Early Hints Callback Request","operationId":"connect#PesaLinkSendToPhone","parameters":[{"name":"PesaLinkSendToPhoneRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/ConnectPesaLinkSendToPhoneRequestBody","required":["MessageReference","MessageDateTime","MessageCode","MessageDescription","Source","Destinations"]}}],"responses":{"200":{"description":"OK response.","schema":{"type":"string"}},"207":{"description":"Multi-Status response.","schema":{"$ref":"#/definitions/ConnectPesaLinkSendToPhoneFullFailureResponseBody"}}},"schemes":["http"]}},"/callbacks/coop-bank/FundsTransfer/Internal/A2A/2.0.0":{"post":{"tags":["connect"],"summary":"IFTAccountToAccount connect","description":"IFTAccountToAccount Early Hints Callback Request","operationId":"connect#IFTAccountToAccount","parameters":[{"name":"IFTAccountToAccountRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/ConnectIFTAccountToAccountRequestBody","required":["MessageReference","MessageDateTime","MessageCode","MessageDescription","Source","Destinations"]}}],"responses":{"200":{"description":"OK response.","schema":{"type":"string"}},"207":{"description":"Multi-Status response.","schema":{"$ref":"#/definitions/ConnectIFTAccountToAccountFullFailureResponseBody"}}},"schemes":["http"]}},"/health":{"get":{"tags":["health"],"summary":"show health","description":"Health check endpoint.","operationId":"health#show","produces":["text/plain"],"responses":{"200":{"description":"OK response.","schema":{"type":"string"}}},"schemes":["http"]}},"/swagger/swagger.json":{"get":{"tags":["swagger"],"summary":"Download ../../gen/http/openapi.json","description":"JSON document containing the API swagger definition","operationId":"swagger#/swagger/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}}},"definitions":{"ConnectIFTAccountToAccountFullFailureResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"2 - FULL FAILURE (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"ConnectIFTAccountToAccountPartialSuccessResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"1 - PARTIAL SUCCESS (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"ConnectIFTAccountToAccountRequestBody":{"title":"ConnectIFTAccountToAccountRequestBody","type":"object","properties":{"Destinations":{"$ref":"#/definitions/DestinationAccountCallbackRequestRequestBody"},"MessageCode":{"type":"string","description":"Message Response Code","example":"0"},"MessageDateTime":{"type":"string","description":"Acknowledgement message creation timestamp","example":"2017-12-04T09:27:00","format":"date-time"},"MessageDescription":{"type":"string","description":"Message Code description","example":"FULL SUCCESS"},"MessageReference":{"type":"string","description":"Your unique transaction request message identifier","example":"40ca18c6765086089a1","minLength":1,"maxLength":27},"Source":{"$ref":"#/definitions/SourceAccountCallbackRequestRequestBody"}},"example":{"Destinations":{"Amount":777,"MobileNumber":"2547xxxxxxxx","Narration":"Electricity Payment","ReferenceNumber":"40ca18c6765086089a1_1","ResponseCode":"0","ResponseDescription":"Success","TransactionID":"116e68e0af0c38zY_NKE7HBQEIL"},"MessageCode":"0","MessageDateTime":"2017-12-04T09:27:00","MessageDescription":"FULL SUCCESS","MessageReference":"40ca18c6765086089a1","Source":{"AccountNumber":"54321987654321","Amount":777,"Narration":"Electricity Payment","ResponseCode":"0","ResponseDescription":"Success","TransactionCurrency":"KES"}},"required":["MessageReference","MessageDateTime","MessageCode","MessageDescription","Source","Destinations"]},"ConnectPesaLinkSendToAccountFullFailureResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"2 - FULL FAILURE (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"ConnectPesaLinkSendToAccountPartialSuccessResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"1 - PARTIAL SUCCESS (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"ConnectPesaLinkSendToAccountRequestBody":{"title":"ConnectPesaLinkSendToAccountRequestBody","type":"object","properties":{"CallBackUrl":{"type":"string","description":"Your callback URL that will receive transaction processing results","example":"https://yourdomain.com/ft-callback"},"Destinations":{"$ref":"#/definitions/DestinationAccountCallbackRequestRequestBody"},"MessageCode":{"type":"string","description":"Message Response Code","example":"0"},"MessageDateTime":{"type":"string","description":"Acknowledgement message creation timestamp","example":"2017-12-04T09:27:00","format":"date-time"},"MessageDescription":{"type":"string","description":"Message Code description","example":"FULL SUCCESS"},"MessageReference":{"type":"string","description":"Your unique transaction request message identifier","example":"40ca18c6765086089a1","minLength":1,"maxLength":27},"Source":{"$ref":"#/definitions/SourceAccountCallbackRequestRequestBody"}},"example":{"CallBackUrl":"https://yourdomain.com/ft-callback","Destinations":{"Amount":777,"MobileNumber":"2547xxxxxxxx","Narration":"Electricity Payment","ReferenceNumber":"40ca18c6765086089a1_1","ResponseCode":"0","ResponseDescription":"Success","TransactionID":"116e68e0af0c38zY_NKE7HBQEIL"},"MessageCode":"0","MessageDateTime":"2017-12-04T09:27:00","MessageDescription":"FULL SUCCESS","MessageReference":"40ca18c6765086089a1","Source":{"AccountNumber":"54321987654321","Amount":777,"Narration":"Electricity Payment","ResponseCode":"0","ResponseDescription":"Success","TransactionCurrency":"KES"}},"required":["MessageReference","MessageDateTime","MessageCode","MessageDescription","Source","Destinations"]},"ConnectPesaLinkSendToPhoneFullFailureResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"2 - FULL FAILURE (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"ConnectPesaLinkSendToPhonePartialSuccessResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"1 - PARTIAL SUCCESS (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"ConnectPesaLinkSendToPhoneRequestBody":{"title":"ConnectPesaLinkSendToPhoneRequestBody","type":"object","properties":{"Destinations":{"type":"array","items":{"$ref":"#/definitions/DestinationAccountCallbackRequestRequestBody"},"example":[{"Amount":777,"MobileNumber":"2547xxxxxxxx","Narration":"Electricity Payment","ReferenceNumber":"40ca18c6765086089a1_1","ResponseCode":"0","ResponseDescription":"Success","TransactionID":"116e68e0af0c38zY_NKE7HBQEIL"},{"Amount":777,"MobileNumber":"2547xxxxxxxx","Narration":"Electricity Payment","ReferenceNumber":"40ca18c6765086089a1_1","ResponseCode":"0","ResponseDescription":"Success","TransactionID":"116e68e0af0c38zY_NKE7HBQEIL"}],"minItems":1},"MessageCode":{"type":"string","description":"Message Response Code","example":"0"},"MessageDateTime":{"type":"string","description":"Acknowledgement message creation timestamp","example":"2017-12-04T09:27:00","format":"date-time"},"MessageDescription":{"type":"string","description":"Message Code description","example":"FULL SUCCESS"},"MessageReference":{"type":"string","description":"Your unique transaction request message identifier","example":"40ca18c6765086089a1","minLength":1,"maxLength":27},"Source":{"$ref":"#/definitions/SourceAccountCallbackRequestRequestBody"}},"example":{"Destinations":[{"Amount":777,"MobileNumber":"2547xxxxxxxx","Narration":"Electricity Payment","ReferenceNumber":"40ca18c6765086089a1_1","ResponseCode":"0","ResponseDescription":"Success","TransactionID":"116e68e0af0c38zY_NKE7HBQEIL"},{"Amount":777,"MobileNumber":"2547xxxxxxxx","Narration":"Electricity Payment","ReferenceNumber":"40ca18c6765086089a1_1","ResponseCode":"0","ResponseDescription":"Success","TransactionID":"116e68e0af0c38zY_NKE7HBQEIL"}],"MessageCode":"0","MessageDateTime":"2017-12-04T09:27:00","MessageDescription":"FULL SUCCESS","MessageReference":"40ca18c6765086089a1","Source":{"AccountNumber":"54321987654321","Amount":777,"Narration":"Electricity Payment","ResponseCode":"0","ResponseDescription":"Success","TransactionCurrency":"KES"}},"required":["MessageReference","MessageDateTime","MessageCode","MessageDescription","Source","Destinations"]},"ConnectSendToMPesaFullFailureResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"2 - FULL FAILURE (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"ConnectSendToMPesaPartialSuccessResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"1 - PARTIAL SUCCESS (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"ConnectSendToMPesaRequestBody":{"title":"ConnectSendToMPesaRequestBody","type":"object","properties":{"CallBackUrl":{"type":"string","description":"Your callback URL that will receive transaction processing results","example":"https://yourdomain.com/ft-callback"},"Destinations":{"$ref":"#/definitions/DestinationAccountCallbackRequestRequestBody"},"MessageCode":{"type":"string","description":"Message Response Code","example":"0"},"MessageDateTime":{"type":"string","description":"Acknowledgement message creation timestamp","example":"2017-12-04T09:27:00","format":"date-time"},"MessageDescription":{"type":"string","description":"Message Code description","example":"FULL SUCCESS"},"MessageReference":{"type":"string","description":"Your unique transaction request message identifier","example":"40ca18c6765086089a1","minLength":1,"maxLength":27},"Source":{"$ref":"#/definitions/SourceAccountCallbackRequestRequestBody"}},"example":{"CallBackUrl":"https://yourdomain.com/ft-callback","Destinations":{"Amount":777,"MobileNumber":"2547xxxxxxxx","Narration":"Electricity Payment","ReferenceNumber":"40ca18c6765086089a1_1","ResponseCode":"0","ResponseDescription":"Success","TransactionID":"116e68e0af0c38zY_NKE7HBQEIL"},"MessageCode":"0","MessageDateTime":"2017-12-04T09:27:00","MessageDescription":"FULL SUCCESS","MessageReference":"40ca18c6765086089a1","Source":{"AccountNumber":"54321987654321","Amount":777,"Narration":"Electricity Payment","ResponseCode":"0","ResponseDescription":"Success","TransactionCurrency":"KES"}},"required":["MessageReference","MessageDateTime","MessageCode","MessageDescription","Source","Destinations"]},"DestinationAccountCallbackRequestRequestBody":{"title":"DestinationAccountCallbackRequestRequestBody","type":"object","properties":{"Amount":{"type":"integer","description":"Transaction Amount","example":777,"minimum":0,"maximum":999999},"MobileNumber":{"type":"string","description":"Recipient M-Pesa mobile number","example":"2547xxxxxxxx","minLength":10,"maxLength":13},"Narration":{"type":"string","description":"Transaction posting narration","example":"Electricity Payment","minLength":1,"maxLength":25},"ReferenceNumber":{"type":"string","description":"Unique posting reference for the transaction leg","example":"40ca18c6765086089a1_1","minLength":1,"maxLength":30},"ResponseCode":{"type":"string","description":"Posting leg response code","example":"0"},"ResponseDescription":{"type":"string","description":"Posting leg response description","example":"Success"},"TransactionID":{"type":"string","description":"Co-operative Bank's processed transaction number concatenated with M-Pesa transaction number","example":"116e68e0af0c38zY_NKE7HBQEIL"}},"description":"Destination Account Callback Request","example":{"Amount":777,"MobileNumber":"2547xxxxxxxx","Narration":"Electricity Payment","ReferenceNumber":"40ca18c6765086089a1_1","ResponseCode":"0","ResponseDescription":"Success","TransactionID":"116e68e0af0c38zY_NKE7HBQEIL"},"required":["ReferenceNumber","MobileNumber","Amount","Narration"]},"SourceAccountCallbackRequestRequestBody":{"title":"SourceAccountCallbackRequestRequestBody","type":"object","properties":{"AccountNumber":{"type":"string","description":"Posting account number","example":"54321987654321","minLength":14,"maxLength":14},"Amount":{"type":"integer","description":"Transaction Amount","example":777,"minimum":0.01,"maximum":999999.99},"Narration":{"type":"string","description":"Posting account transaction narration","example":"Electricity Payment","minLength":1,"maxLength":25},"ResponseCode":{"type":"string","description":"Posting leg response code","example":"0"},"ResponseDescription":{"type":"string","description":"Posting leg response description","example":"Success"},"TransactionCurrency":{"type":"string","description":"Posting account currency in ISO Currency Code","example":"KES","enum":["USD","KES","EUR","GBP","AUD","CHF","CAD","ZAR"]}},"description":"Source Account Callback Request","example":{"AccountNumber":"54321987654321","Amount":777,"Narration":"Electricity Payment","ResponseCode":"0","ResponseDescription":"Success","TransactionCurrency":"KES"},"required":["AccountNumber","Amount","TransactionCurrency","Narration","ResponseCode","ResponseDescription"]}},"externalDocs":{"description":"Library Usage","url":"https://github.com/wondenge/coop-notifiers/blob/master/README.md"}}